add PageContainer component, move viewport states and functions to app.js, change some margin-styles
