+++ pages/contact.js
@@ -1,24 +1,9 @@
 import styled from "styled-components";
 import { css } from "styled-components";
 import { useState, useEffect } from "react";
+import PageContainer from "@/components/PageContainer";
 
-export default function Contact({ language }) {
-  const [viewportWidth, setViewportWidth] = useState("");
-  const [viewportHeight, setViewportHeight] = useState("");
-
-  function handleResize() {
-    setViewportWidth(window.innerWidth);
-    setViewportHeight(window.innerHeight);
-  }
-
-  useEffect(() => {
-    if (typeof window !== undefined) {
-      setViewportWidth(window.innerWidth);
-      setViewportHeight(window.innerHeight);
-      window.addEventListener("resize", handleResize);
-      return () => window.removeEventListener("resize", handleResize);
-    }
-  }, []);
+export default function Contact({ language, viewportHeight, viewportWidth }) {
   return (
     <>
       <PageTitle isOnMobile={viewportHeight > viewportWidth}>Contact</PageTitle>
@@ -60,20 +45,13 @@ const PageTitle = styled.h1`
         `}
 `;
 
-const PageContainer = styled.main`
-  background: none;
-  position: relative;
-  max-width: 1100px;
-  margin: ${({ isOnMobile }) => (isOnMobile ? "5vh 0" : "25vh auto 0 auto")};
-`;
-
 const Article = styled.article`
   background: none;
   position: relative;
   display: flex;
   flex-direction: column;
   gap: 40px;
-  margin: ${({ isOnMobile }) => (isOnMobile ? "0 10vw" : "0 20vw 30px 5vw")};
+  margin: ${({ isOnMobile }) => (isOnMobile ? "0 10vw" : "0 10vw 30px 2.55vw")};
 `;
 
 const Title = styled.h2`
